requirements.txt
fastapi
uvicorn
python-socketio
spacy
pandas
numpy
joblib
nltk
jinja2
scikit-learn
tensorflow
pillow